23.3. Pagination Context
Let's limit this to showing 7 results per page. To do that, go over to DailyStats and pass a new option: paginationItemsPerPage set to 7
This is why we read the max per page - which is the $limit - from API Platform's Pagination object instead of hardcoding it.
Now when we move over and try it... beautiful! Seven items and five pages total! Say hello to our home-rolled pagination!
Next: we have our get item and get collection operations working - even with pagination! Let's see if we can get the put operation to work so that we can update DailyStats data.