24.1. Custom Resource PUT Operation
Back to the API documentation! Let's pretend that we also need to be able to update a DailyStats. Maybe, if you're an admin, sometimes you need to double-check the data and update it. Specifically, let's make it possible to change the totalVisitors field.
Ok, cool! Updating is done via the put operation. Over in DailyStats, under itemOperations, add put
Then, because this is the first time we will be denormalizing, copy normalizationContext, paste, rename it to denormalizationContext and set its groups to daily-stats:write
Take that daily-stats:write group and put it above totalVisitors
Let's give it a try! No, it won't work yet but... it will kind of seem like it's working. Refresh the docs. There's our PUT operation. Execute the collection operation so we can get a valid ID. Perfect. I'll copy this 2020-09-03. Down in the put operation... actually, let me scroll back up so we can see that the totalVisitors is currently 1,500.
Down on the put operation, hit "Try it out", paste the date string as the id and set totalVisitors to 500. Hit \Execute" and... it works! Wait, it worked?
No, it didn't really work. The deserialization process did update the DailyStats object, which is why we see the correct number in the response. But it's not actually saving. If you re-tried the get collection operation... yep! It's still 1,500.