25.2. Custom Filter, getDescription() & properties
Now, in our case, we're creating this filter entirely for a single class for our application. So we don't need to worry about making it configurable. Remove the properties option from the annotation
We're going to completely ignore $this->properties... and do whatever we want.
Next, we now know enough to fill in getDescription() for our filter. Once we've done that, we'll bring it to life!