26.1. Custom Filter Logic for Entities
Let's review our goal: to be able to add ?search=cheese and have that search across several columns in the database.
In getDescription(), let's describe that lofty goal. We don't need any fancy, dynamic property stuff like we saw in the core filters. Just return an array with one item for the one query parameter: search. Set that to another array and... here is where we start using the keys that we saw inside PropertyFilter
I'll cheat and list the ones we need. First, set 'property' => null
If a filter does relate to a specific property, put that here. As far as I can tell, this is only used in the Hydra filter documentation, not even in OpenAPI, which is what drives the Swagger interactive docs. Wow, that sentence was FULL of buzzwords. Phew!
Next set 'type' => 'string' and 'required' => false
Both things that will help the docs.
Let's check it out! Find the API docs, refresh and open up the /api/cheeses operation. And... there it is: search! And it says "string".
We can even help this a bit more. Add another key called openapi. Set that to another array. One of the keys you can pass here is called description. How about:

  search across multiple fields.
I personally do not know all the possible keys that we can have... I've just been digging around and finding what I need. So feel free to dig further.
When we refresh now and go back to that operation... nice! Our filter now has a description!

How & When filterProperty() is Called
Okay, enough with the documentation: let's get to the part where we actually apply that filter. For a Doctrine ORM filter, this happens in filterProperty(), which receives $property and $value arguments and then some Doctrine-specific stuff like QueryBuilder, something called a QueryNameGenerator - more on that in a few minutes - and a some other stuff
Let's figure out what that $property and $value are: dd() both of these
Then go find the other tab where we're fetching the cheese collection and hit enter to send the ?search=cheese query param.
Cool! It passes us search - which is the name of the query parameter - and also cheese. Now, the really important thing to understand is that filterProperty() is going to be called for every single query parameter that's on the URL. So if I go up here and say ?dog=bark, it prints out dog and bark. Now, obviously our filter is not meant to operate on a dog, we'll leave that to trained