26.2. Custom Filter Logic for Entities
Back in filterProperty(), check for our query parameter: if $property does not equal 'search', then return

Modifying the Query
The rest of this method is good-old query-building logic. We're passed the QueryBuilder that will be used to fetch the collection, and our job is to modify it.
To do that, we first need to know the class alias that's being used for the query. We can get that by saying $alias = $queryBuilder->getRootAliases() and - it looks a bit funny - but we want the 0 index
Now add the real logic: $queryBuilder->andWhere(), pass this sprintf() - because the alias will be dynamic - and search on both the title and description fields. So, %s.title LIKE :search OR %s.description LIKE :search. For the 2 %s, pass $alias and $alias. And... I'll split this onto multiple lines
Finish with ->setParameter() to assign the search parameter to '%'..'%' for a fuzzy search
Let's try this! First, remove the query parameter entirely so we can see what the full list looks like. Ok, a lot of people are selling blocks of cheddar. Add ?search=cheddar
/api/cheeses/1 is gone but the rest do have cheddar in their title.
Let's try the word "cube" to see if the description is matching. And... that works too!
To prove it, we can even see the query. Go to /_profiler... click the token for our API request, click into the Doctrine section then "view formatted query". Beautiful! The is_published and owner_id check comes from a Doctrine extension we created in the last tutorial and relates to security. And then it searches on the title or description fields. Pretty cool.