26.3. Custom Filter Logic for Entities
The QueryNameGenerator
Before we keep going, one argument we did not use was .
The query name generator is probably not very important unless you're creating a filter that you want to share between projects.
Here's the problem it solves: the parameter we added - search - could have been called anything - it just needs to match the :search inside the query
Now, if there are many independent filters being used, then, in theory, two filters might accidentally use the same parameter name. If that happened one would override the other. That's no fun.
The query name generator's job is to help avoid this problem by generating a unique parameter name.
Check it out: say $valueParameter = $queryNameGenerator-> - that's the argument we're being passed - then generateParameterName('search')
That will return a string with search then a unique index that increments: something like search_p1 or search_p2. I'll put a comment above this
Down in the query, using it does get ugly: instead of :search, it's :%s and then another :%s. For the arguments, we need $alias, $valueParameter, $alias, $valueParameter
Finally, in setParameter(), use $valueParameter
I'll be honest, that makes my head spin a little... and I might avoid doing this for custom filters in my own project.
Anyways, let's make sure it works by going to the documentation. Hit "Try it Out", fill in "cube" for the search and... Execute! Let's see... yep! Our filter is working!
So this is what it looks like to make a custom filter when your resource is a Doctrine entity. But the process is different if you need to make a custom filter for an API resource that is not an entity. Let's tackle that next by making a filter for DailyStats.