27.1. Custom Filter for Custom Resources
Let's create one more custom filter: this time for DailyStats. Here's our mission: to be able to add a ?from= date on the URL. For example, if we go to /api/daily-stats.jsonld, we see all the stats. But if we add ?from=2020-09-01, we want to only show stats on that day or later.
Let's do this! Start basically the same way as before: in the src/ApiPlatform/ directory, create a new PHP class called DailyStatsDateFilter

Creating the Filter Class
For the search filter, we extended AbstractFilter
but that class only works for entity resources. This time, implement FilterInterface. Now, be careful. There are multiple FilterInterface in API platform. One is for the ORM, and even inside the main part of API Platform, there are 2: one in Serializer and one in Api. Get the one in Serializer
This FilterInterface actually extends the other one from Api\ but it has extra integration that will make our life much easier.
Back in the new class, go to "Code"->"Generate" - or Command + N on a Mac - and hit "Implement Methods" to add the two we need... which are quite similar to the two methods we had before
The only difference is that the apply() method has different arguments than filterProperty()... which makes sense because that method was all about querying via Doctrine.
Before we start filling this in, go to src/Entity/DailyStats.php so we can use the filter. So @ApiFilter(), hit tab to auto-complete that - and DailyStatsDateFilter::class... but this time we need to manually add the use statement: use DailyStatsDateFilter
Now that we've activated the filter for this resource, let's jump straight into getDescription(). This will be exactly like what we did in CheeseSearchFilter. In fact, let's go steal our code! Copy the return statement from the other filter... and paste it here. This time the query parameter will be called from, keep 'property' => null, type string is good, required false is good, but tweak the description:

  From date e.g. 2020-09-01
The apply() method is still empty, but this should be enough to see the filter in the docs. Refresh that page, open the DailyStats collection operation and... there it is!