27.2. Custom Filter for Custom Resources
Adding the Filter Logic
But... how can we make this filter actually work? Well, forget about the filter system for a minute. No matter what fanciness we do, it's 100% up to our data provider to return the collection of data that will be shown.
What I mean is, we're not going to return this DailyStatsPaginator object and then expect some other system to somehow magically filter those results
Nope, if we want DailyStats to have some filters, we need to handle that logic inside getCollection() so that the items inside the paginator represent the filtered items.
So before we really think about reading query parameters or integrating with our filter class, let's first focus on making our DailyStatsProvider able to filter the results by date.
And actually, most of the work of figuring out which results to return is done by the paginator. So let's jump into DailyStatsPaginator.
Down here, getIterator() is where we use StatsHelper to fetch all of the objects
And, it already has a type of filtering: it returns a subset of the items based on what page we're on.
And fortunately, if we jump into src/Service/StatsHelper.php, the fetchMany() method already has a third argument called , which supports from and to keys, which are both DateTime objects
Basically, we can already pass a from key on the 3rd argument to filter exactly like we want. The boring work was done for us. Yay!
Ok, let's think: in order for DailyStatsPaginator to be able to return the correct results, it needs to know what the from date is. To store that info, add a property: a new private $fromDate. Above this, I'll document that this will be a \DateTimeInterface or null if it's not set
Then, at the bottom, go to "Code"->"Generate" - or Command+N on a Mac - select setters, and generate the setFromDate() method. Oh, but I'll remove the nullable type
If you call this method... I'll assume that you do have a DateTime that you want to use. We'll see who sets this in a minute.
But assuming that this is set, we can use it inside of getIterator(). Add $criteria equals an empty array. Then, if we have $this->fromDate, say $criteria['from'] = $this->fromDate
Finally, pass $criteria as that third argument to fetchMany()
This class is ready! Let's make sure it's working by going to DailyStatsProvider and hard-coding a date temporarily. Let's see, remove the return statement, add $paginator = and, at the bottom return $paginator. In between add: $paginator->setFromDate(), with a new \DateTime() and 2020-08-30
Let's see if it works! Back at the browser, I'll go over to my other tab. I do have a ?from=, but that query parameter is not being used yet: the filtering should use the hardcoded date.
When we refresh... let's see. Yes! It returned only 5 results starting with the hard-coded 08-30! Awesome!
Oh, but I do see one, teenie-tiny bug: it says hydra:totalItems 30. That should only say 5.
This number comes from our paginator object: the getTotalItems() method, which calls $this->statsHelper->count()
Really, the count() method should also allow an array of $criteria to be passed so that it can do the same filtering that happens in fetchMany(). But... I'll leave that for you as extra credit.
Next: this is good! We have the filtering logic working. But how can we fetch the real data from the query parameter? Well... I sort of just answered my own question! We could grab the Request object and read the query parameter directly. But there's actually a cooler, more built-in way that involves our DailyStatsDateFilter class.