28.1. Custom Filter apply()
Let's talk about how filter classes work internally. As we know, each data provider is 100% responsible for taking filters into account and changing the data it returns based on them. So, filtering happens inside each data provider, not via some magic system that runs after them.

How Filters work in Doctrine
Let's look at how this is done in the core Doctrine data provider. Hit Shift+Shift, search for doctrinedataprovider and include non project items. There it is: CollectionDataProvider from Orm\. Here is the getCollection() method.
The Doctrine data provider has a system called "collection extensions": these are hook points that allow you to modify the query in any way you want. And... we actually created one of these extensions in the last tutorial: CheeseListingIsPublishedExtension:

 57 lines  src/ApiPlatform/CheeseListingIsPublishedExtension.php
class CheeseListingIsPublishedExtension implements QueryCollectionExtensionInterface, QueryItemExtensionInterface
{
    private $security;
    public function __construct(Security $security)
    {
        $this->security = $security;
    }
    public function applyToCollection(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null)
    {
        $this->addWhere($queryBuilder, $resourceClass);
    }
    public function applyToItem(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, array $identifiers, string $operationName = null, array $context = [])
    {
        $this->addWhere($queryBuilder, $resourceClass);
    }
    private function addWhere(QueryBuilder $queryBuilder, string $resourceClass): void
    {
        if ($resourceClass !== CheeseListing::class) {
            return;
        }
        if ($this->security->isGranted('ROLE_ADMIN')) {
            return;
        }
        $rootAlias = $queryBuilder->getRootAliases()[0];
        if (! $this->security->getUser()) {
            $queryBuilder->andWhere(sprintf('%s.isPublished = :isPublished', $rootAlias))
                ->setParameter('isPublished', true);
        } else {
            $queryBuilder->andWhere(sprintf('
                    %s.isPublished = :isPublished
                    OR %s.owner = :owner',
                $rootAlias, $rootAlias
            ))
                ->setParameter('isPublished', true)
                ->setParameter('owner', ->security->getUser());
        }
    }
}
This modifies the query so that we don't return unpublished listings, unless you're the owner of the listing or an admin.

Doctrine's FilterExtension
Why are we talking about these extension classes? Because one of the core Doctrine extensions is called FilterExtension.
Let's open it up: Shirt+Shift and look for FilterExtension.php making sure to include all non-project items. Get the one from Orm\. I love this. It loops over all of the filters that have been activated for this resource class, calls apply() on each one and passes it the QueryBuilder!
Thanks to this, in CheeseSearchFilter, all we needed to do was extend AbstractFilter and fill in the filterProperty() method:

 38 lines  src/ApiPlatform/CheeseSearchFilter.php
class CheeseSearchFilter extends AbstractFilter
{
    protected function filterProperty(string $property, $value, QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null)
    {
    }
}
The apply() method lives in AbstractFilter, which does some work and then ultimately calls filterProperty().
The point is: the Doctrine filter system works via a Doctrine extension, which knows to call a method on each filter object.

How / When the apply() Method is Called
But all of this stuff will not happen in our situation... because we are not using the Doctrine data provider. However, because we made our filter implement the FilterInterface from the core Serializer\ namespace, API Platform will help us a bit:

 28 lines  src/ApiPlatform/DailyStatsDateFilter.php
use ApiPlatform\Core\Serializer\Filter\FilterInterface;
class DailyStatsDateFilter implements FilterInterface
{
}
How? By automatically calling our apply() method on every request for an API resource where our filter has been activated:

 28 lines  src/ApiPlatform/DailyStatsDateFilter.php
class DailyStatsDateFilter implements FilterInterface
{
    public function apply(Request $request, bool $normalization, array $attributes, array &$context)
    {
    }
}
What I mean is, in DailyStats we added @ApiFilter(DailyStatsDateFilter::class):

 60 lines  src/Entity/DailyStats.php
/**
 * @ApiFilter(DailyStatsDateFilter::class)
 */
class DailyStats
{
}
Thanks to this, whenever we make a request to a DailyStats operation, API Platform will automatically call the DailyStatsDateFilter::apply() method:

 28 lines  src/ApiPlatform/DailyStatsDateFilter.php
class DailyStatsDateFilter implements FilterInterface
{
    public function apply(Request $request, bool $normalization, array $attributes, array &$context)
    {
    }
}
This works via a context builder in the core of API Platform that loops over all of the filters for the current resource class, checks to see if they implement the FilterInterface that we're using and, if they do, calls apply().
So... this is huge! It means that API Platform is smart enough to automatically call our filter's apply() method but only when needed. This means that we can get down to work.

Grabbing the Query Parameter
Our first job is to read the query parameter from the URL. And... hey! We get the Request object as an argument
Schweet! Let's dd($request->query->all())
Back at your browser, refresh and... there it is: the from query param.