28.3. Custom Filter apply()
Passing Info from the Filter to the Data Provider
So... what do we do with that string? I mean, we're not inside DailyStatsProvider where we actually need this info: we're way over here in the filter class.
The answer is that we're going top pass this info from the filter to the data provider via the $context. Check it out: one of the arguments to apply() is the  array and it's passed by reference:

 35 lines  src/ApiPlatform/DailyStatsDateFilter.php
class DailyStatsDateFilter implements FilterInterface
{
    public function apply(Request $request, bool $normalization, array $attributes, array &$context)
    {
    }
}
That means we can modify it.
Head to the top of this class and add a new public constant, how about: FROM_FILTER_CONTEXT set to daily_stats_from. This will be the key we set on $context
Before we do that, let's convert the string into a DateTime object: $fromDate = \DateTimeImmutable::createFromFormat() passing Y-m-d as the format and then $from
We're using createFromFormat() because if the $from string is in an invalid format, it will return false. We can use that to code defensively: if $fromDate, then we know we have a valid date. Also add $fromDate = $fromDate->setTime() and pass zero, zero, zero to normalize all the dates to midnight
Finally, set this on the context: $context[self::FROM_FILTER_CONTEXT] = $fromDate
So the job of the apply() method in a custom, non-Doctrine filter is not actually to apply the filtering logic: it's to pass some filtering info into the context.