28.4. Custom Filter apply()
Reading the Context in the Data Provider
And now, we're dangerous. Well... we're almost dangerous. If we can get access to the  from inside DailyStatsProvider, then we can read that key off and set the from date. Unfortunately, we do not have the context yet
But fortunately, we know how to get it!
Instead of CollectionDataProviderInterface, implement ContextAwareCollectionDataProviderInterface
The only difference is that getCollection() now has an extra array $context = [] argument
To start, let's dd($context) and see if the filter info is there
Ok, refresh. And... we got it! The daily_stats_from is there! And if we take the from query param off, it still works, but the key is gone.