29.2. Filter Class Arguments
ApiFilter arguments Option
Let's do an extra challenge. Pretend that we want to make this behavior - whether to throw a 400 error on an invalid date format - something that is configurable when we activate the filter.
Okay, this may not be needed unless you're building a reusable filter, but it will reveal some cool stuff about how filters work.
The @ApiFilter() annotation has several options that we can pass to it
Hold Command or Control and click to jump into that core annotation class. Yep! All of these public properties are options that we can technically pass to this annotation. But for the purposes of building a custom filter, the only options that really matter are arguments and properties. We'll talk about properties later.
Close that class. Try this: add arguments={} and then pass a new argument called throwOnInvalid set to true
What does this do? I don't know! Let's refresh and see what happens. Ah, error!

  Class DailyStatsDateFilter does not have argument $throwOnInvalid