29.3. Filter Class Arguments
arguments Option Maps to Constructor Arguments
API platform does a cool, but kind of strange thing: if you pass an arguments option to a filter, it tries to pass that argument - by name - to the constructor of your filter.
Check it out: in DailyStatsDateFilter, add a constructor: public function __construct() with bool and then copy the name of the argument and paste it: $throwOnInvalid. Default this to false in case someone uses the filter without that option
Next, hit Alt+Enter and go to "Initialize properties" to create that property and set it
Finally, in the if statement, add if not $fromDate and $this->throwOnInvalid, then we want to throw that exception
Let's try it! Move back over, refresh and.... got it! We're back to the 400 status code.

The properties Option
So it's kind of weird, but any arguments on the annotation map to constructor arguments by name. Oh, and the one other option that we could pass to the annotation is properties={} if you want to configure that this is supposed to use a certain set of properties. And if you ever put @ApiFilter above a property instead of on top of the class, this properties option is automatically set for you.
Either way, if the properties option is set, it's passed to your filter as an argument called $properties.
So... cool! We now know how we can pass configuration to a filter. But there's more going on than it seems. Next: we'll reveal something that will make our filters a lot more powerful.