30.1. Filter Autowiring
We just learned that if you pass an arguments option to your annotation:

 60 lines  src/Entity/DailyStats.php
/**
 * @ApiFilter(DailyStatsDateFilter::class, arguments={"throwOnInvalid"=true})
 */
class DailyStats
{
}
Then the keys inside get mapped as arguments to the constructor of your filter class
But there's more going on than it seems. Obviously, someone instantiates our filter object... and it's a pretty good guess that API Platform does it directly and uses the arguments option to figure out what arguments it should pass to the constructor.

Out Filter is a Service!
But actually, our filter is a service in the container! We didn't register it directly, API Platform did thanks to our annotation. Each time it sees an ApiFilter annotation, it registers a unique service for that filter class.
Then, after registering it as a service, it takes the arguments option and sets those as named arguments on that service. Why do we care about how our filter objects are instantiated? Because when API Platform registers the service, it also sets it to autowire: true. Yep, this means we can access services in our filter class just like we normally would!
Check it out: add a LoggerInterface $logger argument
I'm adding it as the first argument just to, kind of, prove that the order doesn't matter. Create the $logger property and then add $this->logger = $logger
Now, down in apply(), we can say $this->logger->info(sprintf()), Filtering from date "%s" and pass $from
Let's see if it works! Move over and go back to use a real from date. Refresh and... ok! No error! Now open a new tab and go to /_profiler. Find the 200 status, click the token link and go down to Logs. Got it!
Oh, but it logged twice? That's actually ok. The system that calls the apply() method on our filter is the "serialization context builder" system, a system we hooked into before. In the src/Serializer/ directory, in the last tutorial, we created an AdminGroupsContextBuilder.
Anyways, the context builder system is called two times: once at the beginning when it's reading the data and again later when the data is serialized. Hence, we see two logs