30.2. Filter Autowiring
Adding arguments in an Entity Filter
Anyways, all of this stuff about the arguments option and autowiring applies equally to an entity filter like our CheeseSearchFilter.

For example, inside of this class, we use LIKE in our query instead of equals:

 38 lines  src/ApiPlatform/CheeseSearchFilter.php
class CheeseSearchFilter extends AbstractFilter
{
    protected function filterProperty(string $property, $value, QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null)
    {
        $queryBuilder->andWhere(sprintf('%s.title LIKE :%s OR %s.description LIKE :%s', $alias, $valueParameter, $alias, $valueParameter))
            ->setParameter($valueParameter, '%'..'%');
    }
}
Let's pretend that we want to make this configurable: you can decide if you want a fuzzy or exact search.
Open up src/Entity/CheeseListing.php and find that filter. Let's add arguments={} and invent a new one called useLike set to true
Then, over in a browser, close the profiler and head to /api/cheeses.jsonld. Yep, we immediately get the error we expect:

  CheeseSearchFilter does not have argument $useLike