30.3. Filter Autowiring
Let's go add it! The only weird part in an entity filter is that our parent class already has a constructor. This means we can't just add a new constructor, we need to override the existing one.
To do that, go to "Code"->"Generate" - or Command + N on a Mac - select "Override Methods" and choose __construct()