30.4. Filter Autowiring
Oh, wow: that is a big constructor... which is fine. But we can slim it down a bit. We do need ManagerRegistry, we do need RequestStack, but we don't need to pass the $logger to the parent class: it's not even used there. And you only need the $properties argument if you actually need the properties option on the annotation, or if you allow the filter to be used on a property. Since we don't have this use-case, we don't need it
Ok! Our constructor now looks a bit nicer. For the parent construct call, pass null for the logger, and null for $properties
If we go right now and refresh... we get the same error... but now we can add that argument.